Correct substitute behavior on reallocation and add tests
This fixes the following issue in replace_impl: the call to
try_reserve passed in a difference of capacities, but try_reserve
expects a difference between the desired capacity and the length.
Because the initial capacity was nonzero but the length was zero, this
caused us to reserve less capacity than we ought to have, leading to an
OOB write.

Fix this by reworking replace_impl to have less unsafe code. Now we zero
initialize the buffer, but we also prefer a stack buffer so we may save
an allocation - probably a wash overall.

Add a test for this case.