Match upstream PCRE2_SYS_STATIC env variable behavior
Upstream pcre2-sys crate has tri-state behavior for the PCRE2_SYS_STATIC
environment variable, where 0 is force dynamic linking, 1 is force static
linking, otherwise it is based off the heuristic.

We *do* have the STATIC_PCRE2 feature that upstream doesn't have, but the
benefit of an environment variable is that a) it can be set for transitive
dependencies, b) it can be set without modifying any code or even locally
checking out the project.

This does effectively give PCRE2_SYS_STATIC=0 a veto over the STATIC_PCRE2
feature, but that's probably preferable since it gives maximum control to the
person ultimately building the downstream app (i.e. fish).